Mises à jours tenant compte des derniers retours mineurs et ajout des loaders sur les tableaux.

+++ database/migrations/2020_03_09_084519_create_campagnes_table.php
@@ -17,7 +17,7 @@ class CreateCampagnesTable extends Migration
             $table->id();
             $table->string('nom');
             $table->string('type_remuneration');
-            $table->decimal('remuneration', 16, 8);
+            $table->float('remuneration', 8, 2);
             $table->foreignId('annonceur_id');
             $table->unsignedBigInteger('cree_par');
             $table->unsignedBigInteger('modifie_par')->nullable();
